move `remote_path` into `util`
