+++ Cargo.toml
@@ -121,7 +121,6 @@ members = [
     "crates/refineable/derive_refineable",
     "crates/release_channel",
     "crates/remote",
-    "crates/remote_path",
     "crates/remote_server",
     "crates/repl",
     "crates/reqwest_client",
@@ -336,7 +335,6 @@ recent_projects = { path = "crates/recent_projects" }
 refineable = { path = "crates/refineable" }
 release_channel = { path = "crates/release_channel" }
 remote = { path = "crates/remote" }
-remote_path = { path = "crates/remote_path" }
 remote_server = { path = "crates/remote_server" }
 repl = { path = "crates/repl" }
 reqwest_client = { path = "crates/reqwest_client" }
