--- crates/remote_path/LICENSE-GPL
@@ -1,1 +1,0 @@
-../../LICENSE-GPL
